refactor contribution page and add new component
moved old contributions to contributions_v1 directory, added layout for new contribution page, added SingleStatPanel.tsx
